Adicionandas informações de aluno na nova estrutura
